+++ src/clj_p4/schemas.clj
@@ -232,6 +232,8 @@
    [:exclude-binaries? {:optional true} [:maybe boolean?]]
    [:exclude-categories {:optional true}
     [:maybe [:or [:= :all] [:set keyword?]]]]
+   [:extra-excludes {:optional true} [:maybe [:sequential string?]]]
+   [:includes {:optional true} [:maybe [:sequential string?]]]
    [:fetch-parallelism {:optional true} [:maybe pos-int?]]
    [:max-print-bytes {:optional true} [:maybe pos-int?]]
    [:user-map {:optional true} [:maybe [:map-of string? user-map-entry]]]